requirements.txt
# Core dependencies
flask==3.0.0
python-dotenv==1.0.0
websockets==12.0
requests==2.31.0
openai==1.12.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1
